+++ src/components/network-cost-trend-chart.tsx
@@ -62,8 +62,14 @@ const LatestYearChartView = (
     operatorsNames: Set<string>;
   }
 ) => {
-  const { observations, networkCosts, id, operatorLabel, operatorsNames } =
-    props;
+  const {
+    observations,
+    networkCosts,
+    id,
+    operatorLabel,
+    operatorsNames,
+    compareWith,
+  } = props;
 
   return (
     <DotPlot
@@ -84,7 +90,9 @@ const LatestYearChartView = (
         y: { componentIri: "network_level" },
         segment: {
           componentIri: "operator_name",
-          palette: "elcom",
+          palette: compareWith?.includes("sunshine.select-all")
+            ? "elcom-categorical-3" //Only green hover if all operators are selected
+            : "elcom2", //Corresponding color palette for the tiles inside the selector if not all operators are selected
         },
         style: {
           entity: "operator_id",
@@ -139,20 +147,22 @@ const LatestYearChartView = (
           symbol={"diamond"}
         />
 
-        <LegendItem
-          item={t({
-            id: "network-cost-trend-chart.legend-item.other-operators",
-            message: "Other operators",
-          })}
-          color={palette.monochrome[200]}
-          symbol={"circle"}
-        />
+        {compareWith?.includes("sunshine.select-all") && (
+          <LegendItem
+            item={t({
+              id: "network-cost-trend-chart.legend-item.other-operators",
+              message: "Other operators",
+            })}
+            color={palette.monochrome[200]}
+            symbol={"circle"}
+          />
+        )}
       </Box>
       <ChartContainer>
         <ChartSvg>
           <AxisWidthLinear position="top" format="number" />
           <AxisHeightCategories stretch />
-          <Dots />
+          <Dots compareWith={compareWith} />
           <InteractionDotted />
           <DotPlotMedian />
         </ChartSvg>